+++ crosscompute/routines/server.py
@@ -1,9 +1,8 @@
-from invisibleroads_macros_disk import is_path_in_folder
 from logging import getLogger, DEBUG, ERROR
-from pathlib import Path
+from time import time
+
 from pyramid.config import Configurator
 from pyramid.events import NewResponse
-from time import time
 from waitress import serve
 from watchgod import watch
 
@@ -11,10 +10,7 @@ from ..constants import (
     HOST,
     MAXIMUM_PING_INTERVAL_IN_SECONDS,
     MINIMUM_PING_INTERVAL_IN_SECONDS,
-    MODE_CODE_BY_NAME,
-    MODE_ROUTE,
-    PORT,
-    RUN_ROUTE)
+    PORT)
 from ..exceptions import (
     CrossComputeError)
 from ..macros.process import LoggableProcess, StoppableProcess
@@ -22,6 +18,7 @@ from ..routes.authorization import AuthorizationRoutes
 from ..routes.automation import AutomationRoutes
 from ..routes.mutation import MutationRoutes
 from ..routines.authorization import AuthorizationGuard
+from .database import DiskDatabase
 from .interface import Server
 
 
@@ -64,31 +61,21 @@ class DiskServer(Server):
             L.error(e)
 
     def watch(
-            self,
-            configuration,
-            disk_poll_in_milliseconds,
-            disk_debounce_in_milliseconds,
-            reload):
+            self, configuration, disk_poll_in_milliseconds,
+            disk_debounce_in_milliseconds, reload):
         if L.getEffectiveLevel() > DEBUG:
             getLogger('watchgod.watcher').setLevel(ERROR)
-        server_process, info_by_path = self._serve(configuration)
+        server_process, disk_database = self._serve(configuration)
         automation_folder = configuration.folder
         for changes in watch(
-                automation_folder,
-                min_sleep=disk_poll_in_milliseconds,
+                automation_folder, min_sleep=disk_poll_in_milliseconds,
                 debounce=disk_debounce_in_milliseconds):
+            changed_paths = [_[1] for _ in changes]
+            changed_infos = disk_database.grok(changed_paths)
             should_restart_server = False
-            changed_infos = []
-            for changed_type, changed_path in changes:
-                try:
-                    changed_info = _get_info(
-                        configuration, changed_path, info_by_path)
-                except KeyError:
-                    continue
-                if changed_info['code'] == 'c':
+            for info in changed_infos:
+                if info['code'] == 'c':
                     should_restart_server = True
-                changed_infos.append(changed_info)
-                L.debug('%s %s %s', changed_type, changed_path, changed_info)
             if should_restart_server:
                 try:
                     configuration = reload()
@@ -96,16 +83,14 @@ class DiskServer(Server):
                     L.error(e)
                     continue
                 server_process.stop()
-                server_process, info_by_path = self._serve(configuration)
-            if changed_infos:
-                self._infos_by_timestamp[time()] = changed_infos
+                server_process, disk_database = self._serve(configuration)
 
     def _serve(self, configuration):
         server_process = StoppableProcess(
             name='server', target=self.serve, args=(configuration,))
         server_process.start()
-        info_by_path = _get_info_by_path(configuration)
-        return server_process, info_by_path
+        disk_database = DiskDatabase(configuration, self._infos_by_timestamp)
+        return server_process, disk_database
 
 
 def _get_app(
@@ -197,139 +182,6 @@ def _configure_allowed_origins(config, allowed_origins):
             'Access-Control-Allow-Origin': origin})
 
     config.add_subscriber(update_cors_headers, NewResponse)
-
-
-def _get_info(configuration, path, info_by_path):
-    path = Path(path)
-    real_path = path.resolve()
-    automation_definitions = configuration.automation_definitions
-    for automation_definition in automation_definitions:
-        automation_folder = automation_definition.folder
-        runs_folder = automation_folder / 'runs'
-        if not is_path_in_folder(path, runs_folder):
-            continue
-        run_id = path.resolve().relative_to(runs_folder).parts[0]
-        run_uri = RUN_ROUTE.format(run_slug=run_id)
-        variable_packs = _get_variable_packs_from_folder(
-            configuration, runs_folder / run_id, run_uri)
-        for path, info in variable_packs:
-            if real_path == path.resolve():
-                return info
-    return info_by_path[real_path]
-
-
-def _get_info_by_path(configuration):
-    'Set c = configuration, v = variable, t = template, s = style'
-    packs = []
-    packs.extend(_get_configuration_packs(configuration))
-    packs.extend(_get_variable_packs(configuration))
-    packs.extend(_get_template_packs(configuration))
-    packs.extend(_get_style_packs(configuration))
-    return {path.resolve(): info for path, info in packs}
-
-
-def _get_configuration_packs(configuration):
-    packs = []
-    # Get automation configuration paths
-    packs.append((configuration.path, {}))
-    for import_configuration in configuration.import_configurations:
-        packs.append((import_configuration['path'], {}))
-    # Get batch configuration paths
-    for automation_definition in configuration.automation_definitions:
-        automation_folder = automation_definition.folder
-        # Use raw batch definitions
-        raw_batch_definitions = automation_definition.get('batches', [])
-        for raw_batch_definition in raw_batch_definitions:
-            batch_configuration = raw_batch_definition.get(
-                'configuration', {})
-            if 'path' not in batch_configuration:
-                continue
-            path = automation_folder / batch_configuration['path']
-            packs.append((path, {}))
-    for path, d in packs:
-        d['code'] = 'c'
-    return packs
-
-
-def _get_variable_packs(configuration):
-    packs = []
-    automation_definitions = configuration.automation_definitions
-    for automation_definition in automation_definitions:
-        automation_folder = automation_definition.folder
-        # Use computed batch definitions
-        for batch_definition in automation_definition.batch_definitions:
-            batch_uri = batch_definition.uri
-            variable_packs = _get_variable_packs_from_folder(
-                configuration,
-                automation_folder / batch_definition.folder,
-                batch_uri)
-            packs.extend(variable_packs)
-    return packs
-
-
-def _get_variable_packs_from_folder(
-        configuration, absolute_batch_folder, batch_uri):
-    packs = []
-    automation_definitions = configuration.automation_definitions
-    for automation_definition in automation_definitions:
-        automation_uri = automation_definition.uri
-        d = automation_definition.variable_definitions_by_mode_name
-        for mode_name, variable_definitions in d.items():
-            mode_code = MODE_CODE_BY_NAME[mode_name]
-            mode_uri = MODE_ROUTE.format(mode_code=mode_code)
-            folder = absolute_batch_folder / mode_name
-            for variable_definition in variable_definitions:
-                variable_id = variable_definition.id
-                info = {
-                    'id': variable_id,
-                    'uri': automation_uri + batch_uri + mode_uri}
-                variable_configuration = variable_definition.configuration
-                if 'path' in variable_configuration:
-                    path = folder / variable_configuration['path']
-                    packs.append((path, info))
-                path = folder / variable_definition.path
-                packs.append((path, info))
-    for path, d in packs:
-        d['code'] = 'v'
-    return packs
-
-
-def _get_template_packs(configuration):
-    packs = []
-    automation_definitions = configuration.automation_definitions
-    for automation_definition in automation_definitions:
-        automation_folder = automation_definition.folder
-        automation_uri = automation_definition.uri
-        d = automation_definition.template_definitions_by_mode_name
-        for mode_name, template_definitions in d.items():
-            for template_definition in template_definitions:
-                if 'path' not in template_definition:
-                    continue
-                path = automation_folder / template_definition.path
-                packs.append((path, {
-                    'mode_name': mode_name, 'uri': automation_uri}))
-        d = automation_definition.template_path_by_id
-        for template_id, path in d.items():
-            packs.append((automation_folder / path, {
-                'id': template_id, 'uri': automation_uri}))
-    for path, d in packs:
-        d['code'] = 't'
-    return packs
-
-
-def _get_style_packs(configuration):
-    packs = []
-    automation_definitions = configuration.automation_definitions
-    for automation_definition in automation_definitions:
-        automation_folder = automation_definition.folder
-        for style_definition in automation_definition.style_definitions:
-            if 'path' not in style_definition:
-                continue
-            path = automation_folder / style_definition['path']
-            packs.append((path, {}))
-    for path, d in packs:
-        d['code'] = 's'
-    return packs
 
 
 L = getLogger(__name__)