+++ CHANGES.md
@@ -2,13 +2,14 @@
 - Start from scratch
 - Restore self-contained server
 - Add views: link, string, number, password, email, text, markdown, image
-- Add views: table
+- Add views: table, frame
 - Move map views into crosscompute-views-map: map-mapbox, map-deck-screengrid
 - Refresh variables without reloading page
+- Run batches with process or thread concurrency
 - Support script.function
 - Support display.pages > configuration.design
-- Support environment > interval for timed updates
-- Run batches with process or thread concurrency
+- Support environment > interval to re-run automation batches periodically
+- Support environment > engine=podman to run automation in a container
 
 # 0.8
 - Start from scratch